+++ Items/Accessories/Rope_Of_Sharing.cs
@@ -5,7 +5,7 @@ using Terraria.ModLoader;
 namespace Origins.Items.Accessories {
 	[AutoloadEquip(EquipType.Waist)]
 	public class Rope_Of_Sharing : ModItem {
-		public static sbyte WaistSlot { get; private set; }
+		public static int WaistSlot { get; private set; }
 		public override void SetStaticDefaults() {
 			// DisplayName.SetDefault("Rope of Sharing");
 			// Tooltip.SetDefault("All stats, afflictions and other effects are shared with another connected player");
